Making sure std containers are cleared before serialization loads data/1d_stencil.xlsx
- flyby warning fix in one of the tests
- flyby change: moving ..._impl functions into namespace detail

+++ tests/unit/serialization/serialization_map.cpp
@@ -102,7 +102,7 @@ void test_vector_as_value()
     std::vector<char> buffer;
     hpx::serialization::output_archive oarchive(buffer);
     std::map<size_t, std::vector<int> > os;
-    for (size_t k = 0; k < 10; ++k)
+    for (int k = 0; k < 10; ++k)
     {
         std::vector<int> vec(10);
         std::iota(vec.begin(), vec.end(), k);
